+++ AmazingChess.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AmazingChess", "src\AmazingChess\AmazingChess.csproj", "{C12F18BD-156A-4EF0-9CF6-5B13200A2C27}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AmazingChessTest", "test\AmazingChessTest\AmazingChessTest.csproj", "{CFAB9505-B4C1-40E7-9145-CB6EE8A51D9B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{C12F18BD-156A-4EF0-9CF6-5B13200A2C27}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C12F18BD-156A-4EF0-9CF6-5B13200A2C27}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C12F18BD-156A-4EF0-9CF6-5B13200A2C27}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CFAB9505-B4C1-40E7-9145-CB6EE8A51D9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CFAB9505-B4C1-40E7-9145-CB6EE8A51D9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CFAB9505-B4C1-40E7-9145-CB6EE8A51D9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CFAB9505-B4C1-40E7-9145-CB6EE8A51D9B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
